:sparkles: Add : 구조변경
