+++ apps/web/src/app/(needLogin)/travel/plan/page.tsx
@@ -1,24 +1,25 @@
-// 'use client';
-// import { useSearchParams } from 'next/navigation';
+'use client';
 
-// import { travelPlanStep1config } from '@/constants';
+import { useSearchParams } from 'next/navigation';
 
-// import Step1Title from '../_components/Step1Title/Step1Title';
-// import TravelPlanTemplete from './stepTemplete';
+import { travelPlanStep1config } from '@/constants';
 
-// const TravelPlan = () => {
-//   const params = useSearchParams();
-//   const search = params.get('stepId');
+import Step1Title from '../_components/Step1Title/Step1Title';
+import TravelPlanTemplete from './travelPlanTemplete';
 
-//   let StepComponent;
+const TravelPlan = () => {
+  const params = useSearchParams();
+  const search = params.get('stepId');
 
-//   if (search == '0') {
-//     StepComponent = <Step1Title config={travelPlanStep1config} />;
-//   } else {
-//     StepComponent = <TravelPlanTemplete search={search} />;
-//   }
+  // `search`가 '0'이면 Step1Title을 사용, 그렇지 않으면 TravelPlanTemplete를 사용
+  const StepComponent =
+    search === '0' ? (
+      <Step1Title config={travelPlanStep1config} />
+    ) : (
+      <TravelPlanTemplete search={search} />
+    );
 
-//   return <div>{StepComponent}</div>;
-// };
+  return <div>{StepComponent}</div>;
+};
 
-// export default TravelPlan;
+export default TravelPlan;
