+++ apps/web/src/app/(needLogin)/travel/record/page.tsx
@@ -1,22 +1,22 @@
 'use client';
+
 import { useSearchParams } from 'next/navigation';
 
 import { travelRecordStep1config } from '@/constants';
 
 import Step1Title from '../_components/Step1Title/Step1Title';
-import TravelRecordTemplete from './stepTemplate';
+import TravelRecordTemplete from './travelRecordTemplete';
 
 const TravelPlan = () => {
   const params = useSearchParams();
   const search = params.get('stepId');
 
-  let StepComponent;
-
-  if (search == '0') {
-    StepComponent = <Step1Title config={travelRecordStep1config} />;
-  } else {
-    StepComponent = <TravelRecordTemplete search={search} />;
-  }
+  const StepComponent =
+    search === '0' ? (
+      <Step1Title config={travelRecordStep1config} />
+    ) : (
+      <TravelRecordTemplete search={search} />
+    );
 
   return <div>{StepComponent}</div>;
 };
