+++ apps/web/src/app/(needLogin)/travel/record/travelRecordTemplete.tsx
@@ -1,3 +1,4 @@
+import { travelRecordStep5config } from '@/constants';
 import { travelRecordStep2config } from '@/constants/travelStep2.constants';
 import { travelRecordStep3config } from '@/constants/travelStep3.constants';
 import { travelRecordStep4config } from '@/constants/travelStep4.constants';
@@ -5,28 +6,20 @@ import { travelRecordStep4config } from '@/constants/travelStep4.constants';
 import Step2When from '../_components/Step2When/Step2When';
 import Step3What from '../_components/Step3What/Step3What';
 import Step4How from '../_components/Step4How/Step4How';
+import Step5Sumup from '../_components/Step5Sumup/Step5Sumup';
 import TravelLayout from '../TravelLayout';
 
+const COMPONENTS_MAP = {
+  '1': <Step2When config={travelRecordStep2config} />,
+  '2': <Step3What config={travelRecordStep3config} />,
+  '3': <Step4How config={travelRecordStep4config} />,
+  '4': <Step5Sumup config={travelRecordStep5config} />,
+};
+
 const TravelRecordTemplete = ({ search }) => {
-  let StepComponent;
+  const StepComponent = COMPONENTS_MAP[search];
 
-  switch (search) {
-    case '1':
-      StepComponent = <Step2When config={travelRecordStep2config} />;
-      break;
-    case '2':
-      StepComponent = <Step3What config={travelRecordStep3config} />;
-
-      break;
-    case '3':
-      StepComponent = <Step4How config={travelRecordStep4config} />;
-      break;
-  }
-  return (
-    <>
-      <TravelLayout>{StepComponent}</TravelLayout>
-    </>
-  );
+  return <TravelLayout>{StepComponent}</TravelLayout>;
 };
 
 export default TravelRecordTemplete;
